text-analysis-report/requirements.txt
pandas
numpy
nltk
spacy
altair
wordcloud
scikit-learn
umap-learn
hdbscan
jinja2
matplotlib
seaborn
scipy
bertopic
sentence-transformers
openai>=1.6.0
python-dotenv
